get tests to work no matter where they are built
This commit fixes a problem with tests failing when the current
directory was different than <root>/SOURCES/build. This was because
the tests where accessing an input file from a directory relative
to the current directory. Hence, if the tests were built and executed
in a random location, the input file couldn't be accessed and the tests
were failing. This behavior was clearly visible in the rpm build
pipeline in jenkins/gerrit.
To solve this, we make cmake define a macro SOURCE_PATH which
contains absolute path to the directory with CMakeLists.txt. Afterwards,
this macro is used in the compiled binaries to navigate to the test
input file.

Change-Id: I22f273a84c873920ae87087de33cd7944931018e