Fix Bias Correction for large numbers
This commit fixes a bug that was rarely causing SIGSEGV
when running HllDistinctCount. Before, for raw estimates
close to the linear counting-bias correction threshold,
the algorithm could access unallocated memory, resulting
in segmentation faults or unpredictable correction values.

Two tests are added as well to test BiasCorrection next to
boundaries of its operation.

This commit bumps the version number to 0.5 to force Jenkins
to rebuild the RPM.

Change-Id: I9e44a0fd47f69ca07f7fc4dc8fe5e04d8a4f79e9